// src/modules/product-adm/usecase/add-product/add-product.dto.ts
export interface AddProductInputDTO {
  id?: string
  name: string
  description: string
  purchasePrice: number
  stock: number
}

export interface AddProductOutputDTO {
  id: string
  name: string
  description: string
  purchasePrice: number
  stock: number
  createdAt: Date
  updatedAt: Date
}